requirements.txt
# Core dependencies
python-telegram-bot==22.2
Flask==3.1.0
flask-cors==4.0.1
google-generativeai==0.5.0
google-cloud-vision==3.10.2
google-api-python-client==2.174.0
APScheduler==3.11.0
python-dotenv==1.0.0
requests==2.32.4
pandas==2.3.0
numpy==2.3.1
matplotlib==3.10.3
seaborn==0.13.2
plotly==5.17.0
Pillow==11.2.1
peewee==3.18.1
psycopg2-binary==2.9.10
weasyprint==62.3
jinja2==3.1.4
beautifulsoup4==4.13.4
feedparser==6.0.11
yfinance==0.2.64
pdf2image==1.17.0
pdfplumber==0.11.7
opencv-python-headless==4.11.0.86
